added check_windows, setup app body, added start_menu, added custom theme in config directory
