+++ src/util/windows.py
@@ -1,0 +1,6 @@
+import platform
+
+def check_for_windows():
+    if platform.system() == 'Windows':
+        from ctypes import windll
+        windll.shcore.SetProcessDpiAwareness(1)
